ALLY-465 (wip) Add new table construction/rendering model
In support of ALLY-465, update the way tables are created.  Prior to
this commit, tables were built and rendered as strings immediately.  For
ALLY-465 (the first introduction of CSV output to the cli) we need to
delay the rendering of tables so that the same data can be used to
render a human readable table, or a CSV.